+++ Corrfunc/theory/DDsmu.py
@@ -24,7 +24,7 @@ def DDsmu(autocorr, nthreads, binfile, mu_max, nmu_bins,
           c_api_timer=False, isa=r'fastest', weight_type=None):
     """
     Calculate the 2-D pair-counts corresponding to the redshift-space
-    correlation function, :math:`\\xi(s, \mu)` Pairs which are separated
+    correlation function, :math:`\\xi(s, \\mu)` Pairs which are separated
     by less than the ``s`` bins (specified in ``binfile``) in 3-D, and
     less than ``s*mu_max`` in the Z-dimension are counted.
 
@@ -34,9 +34,9 @@ def DDsmu(autocorr, nthreads, binfile, mu_max, nmu_bins,
     ``weight_type``.
 
     .. note:: This module only returns pair counts and not the actual
-        correlation function :math:`\\xi(s, \mu)`. See the
+        correlation function :math:`\\xi(s, \\mu)`. See the
         utilities :py:mod:`Corrfunc.utils.convert_3d_counts_to_cf`
-        for computing :math:`\\xi(s, \mu)` from the pair counts.
+        for computing :math:`\\xi(s, \\mu)` from the pair counts.
 
     .. versionadded:: 2.1.0
 
@@ -69,12 +69,12 @@ def DDsmu(autocorr, nthreads, binfile, mu_max, nmu_bins,
         separation from the line of sight (LOS). Here, LOS is taken to be
         along the Z direction.
 
-        Note: Only pairs with :math:`0 <= \cos(\\theta_{LOS}) < \mu_{max}`
+        Note: Only pairs with :math:`0 <= \\cos(\\theta_{LOS}) < \\mu_{max}`
         are counted (no equality).
 
     nmu_bins: int
         The number of linear ``mu`` bins, with the bins ranging from
-        from (0, :math:`\mu_{max}`)
+        from (0, :math:`\\mu_{max}`)
 
     X1/Y1/Z1 : array-like, real (float/double)
         The array of X/Y/Z positions for the first set of points.
@@ -89,7 +89,21 @@ def DDsmu(autocorr, nthreads, binfile, mu_max, nmu_bins,
 
     periodic : boolean
         Boolean flag to indicate periodic boundary conditions.
-        
+
+    boxsize: double or 3-tuple of double, required if ``periodic=True``
+        The (X,Y,Z) side lengths of the spatial domain. Present to facilitate
+        exact calculations for periodic wrapping. A scalar ``boxsize`` will
+        be broadcast to a 3-tuple. If the boxsize in a dimension is 0., then
+        then that dimension's wrap is done based on the extent of the particle
+        distribution. If the boxsize in a dimension is -1., then periodicity
+        is disabled for that dimension.
+
+        .. versionchanged:: 2.4.0
+           Required if ``periodic=True``.
+
+        .. versionchanged:: 2.5.0
+           Accepts a 3-tuple of side lengths.
+
     boxsize : double, required if ``periodic=True``
         The side-length of the cube in the cosmological simulation.
         Present to facilitate exact calculations for periodic wrapping.
@@ -276,6 +290,12 @@ def DDsmu(autocorr, nthreads, binfile, mu_max, nmu_bins,
         msg = "The parameter `mu_max` = {0}, is the max. of cosine of an "\
         "angle and should be within (0.0, 1.0]".format(mu_max)
         raise ValueError(msg)
+
+    if boxsize is not None:
+        boxsize = np.atleast_1d(boxsize)
+        if len(boxsize) == 1:
+            boxsize = (boxsize[0], boxsize[0], boxsize[0])
+        boxsize = tuple(boxsize)
 
     if not autocorr:
         if X2 is None or Y2 is None or Z2 is None: